# src/main/resources/openapi/api.yml
openapi: "3.0.1"
info:
    title: "trip-carrier-advance-api"
    version: "1.0"
    description: "trip-carrier-advance-api"
servers:
    -   url: "{server}/v1"
paths:
    /advance_payment/request_give_advance_payment/{id}:
        post:
            security:
                -   bearerAuth: []
            summary: "Запрос на выдачу аванса"
            operationId: giveAdvanceForTrip
            tags:
                - Advance Payment
            parameters:
                -   name: id
                    in: path
                    description: "ID перевозки"
                    required: true
                    schema:
                        type: integer
                        format: int64
                        minimum: 1
            responses:
                200:
                    description: "OK"
                default:
                    description: "Some other error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                422:
                    description: "Business error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
...
    /advance_payment/is_advanced/{trip_id}:
        get:
            security:
                -   bearerAuth: []
            summary: "Проверить выдан ли аванс"
            operationId: isAdvanced
            tags:
                - Advance Payment
            parameters:
                -   name: trip_id
                    in: path
                    description: "Id заказа на перевозку"
                    required: true
                    schema:
                        type: integer
                        format: int64
                        minimum: 1
            responses:
                200:
                    description: "OK"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/isTripAdvanced"
                default:
                    description: "Some other error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                422:
                    description: "Business error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
...
    /advance_payment/contact/get_contact_carrier/{contrctor_id}:
        get:
            security:
                -   bearerAuth: []
            summary: "Просмотр контактов перевозчика"
            operationId: getContactCarrier
            tags:
                - Cotractor
            parameters:
                -   name: contrctor_id
                    in: path
                    schema:
                        type: integer
                        format: int64
                    required: true
            responses:
                200:
                    description: "OK"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CarrierContactDTO"
                default:
                    description: "Some other error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                422:
                    description: "Business error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
    /advance_payment/contact/update_contact_carrier:
        put:
            security:
                -   bearerAuth: []
            summary: "Обновить контакт контрагента для авансирования"
            operationId: updateContactCarrier
            tags:
                - Cotractor
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CarrierContactDTO"
            responses:
                200:
                    description: "OK"
                default:
                    description: "Some other error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                422:
                    description: "Business error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
    /advance_payment/contact/add_contact_carrier:
        post:
            security:
                -   bearerAuth: []
            summary: "Добавить контакт контрагента для авансирования"
            operationId: addContactCarrier
            tags:
                - Cotractor
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CarrierContactDTO"
            responses:
                200:
                    description: "OK"
                default:
                    description: "Some other error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                422:
                    description: "Business error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
...
    /advance_payment/download_advance_request_template:
        get:
            tags:
                - Application contract
            summary: "Запрос на скачивание шаблонной заявки на аванс"
            operationId: downloadAvanceRequestTemplate
            parameters:
                -   name: trip_num
                    in: query
                    schema:
                        type: string
                    description: "Номер заказа поставщика"
            responses:
                "200":
                    description: "PDF-файл"
                    content:
                        application/pdf:
                            schema:
                                type: string
                                format: binary
                "422":
                    description: "Ошибка в процессе скачивания файла"
...
    /advance_payment/carrier/request_by_carrier/{uuid}:
        post:
            summary: "Хочу аванс экран поставщика"
            operationId: carrierWantsAdvance
            tags:
                - Advance Payment
            parameters:
                -   name: uuid
                    in: path
                    description: "uuid заказа на аванс"
                    required: true
                    schema:
                        type: string
                        format: uuid
            responses:
                200:
                    description: "OK"
                default:
                    description: "Some other error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                422:
                    description: "Business error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
    /advance_payment/carrier/search_request/{uuid}:
        #        без авторизации
        get:
            summary: "Просмотр запроса на аванс для экрана поставщика "
            operationId: searchAdvancePaymentRequestByUuid
            tags:
                - Advance Payment
            parameters:
                -   name: uuid
                    in: path
                    description: "uuid заказа на аванс"
                    required: true
                    schema:
                        type: string
                        format: uuid
            responses:
                200:
                    description: "OK"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CarrierPage'
                default:
                    description: "Some other error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
                422:
                    description: "Business error"
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Error"
...
security:
    -   bearerAuth: []
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
        Error:
            required:
                - status
                - error_code
                - error_message
            properties:
                status:
                    type: string
                error_code:
                    type: string
                error_message:
                    type: string
...
        CarrierPage:
            type: object
            description: "Аванс - Страница перевозчика"
            properties:
                page_carrier_url_is_access:
                    type: boolean
                    description: "Признак доступности ссылки в кабинет перевозчика "
                id:
                    type: integer
                    format: int64
                    description: "ID запроса на авансирование"
                trip_num:
                    type: string
                    description: "Номер перевозки"
                is_cancelled:
                    type: boolean
                    description: "Признак аннулирован аванс"
                trip_cost_with_vat:
                    type: number
                    format: double
                    description: "Стоимость перевозки с НДС"
                advance_payment_sum:
                    type: number
                    format: double
                    description: "Сумма аванса"
                registration_fee:
                    type: number
                    format: double
                    description: "Сбор за оформление документов"
                loading_complete:
                    type: boolean
                    description: "Водитель загрузился"
                loading_address:
                    type: string
                    description: "Адрес первой погрузки"
                unloading_address:
                    type: string
                    description: "Адрес последней разгрузки"
                loading_date:
                    type: string
                    format: date-time
                    description: "Дата погрузки"
                unloading_date:
                    type: string
                    format: date-time
                    description: "Дата разгрузки"
                loading_tz:
                    type: string
                    description: "Таймзона погрузки"
                unloading_tz:
                    type: string
                    description: "Таймзона разгрузки"
#                trip_id:
#                    type: integer
#                    format: int64
#                    description: "ID перевозки"
#                trip_type_code:
#                    type: string
#                    description: "Тип перевозки"
#                created_at:
#                    type: string
#                    format: date-time
#                    description: "Дата создания заказа поставщика"
#                push_button_at:
#                    type: string
#                    format: date-time
#                    description: "Дата и время нажатия на кнопку"
#                author_id:
#                    type: integer
#                    format: int64
#                    description: "Инициатор запроса на авансирование"
#                contractor_id:
#                    type: integer
#                    format: int64
#                    description: "Исполнитель"
#                contractor_name:
#                    type: string
#                    description: "Исполнитель наименование"
#                contact_fio:
#                    type: string
#                    description: "Контакты исполнителя ФИО"
#                contact_email:
#                    type: string
#                    description: "Контакты исполнителя Почта"
#                contact_phone:
#                    type: string
#                    description: "Контакты исполнителя Телефон"
#                payment_contractor:
#                    type: string
#                    description: "От кого везём"
#                req_created_at:
#                    type: string
#                    format: date-time
#                    description: "Время запроса аванса"
#                is_automation_request:
#                    type: boolean
#                    description: "Признак автоматической выдачи аванса"
#                url_contract_application:
#                    type: string
#                    description: "Ссылка на скачивание «Договор-заявка» или «Заявка»"
#                url_advance_application:
#                    type: string
#                    description: "Ссылка на скачивание заявки на авансирование"
#                is_1C_send_allowed:
#                    type: boolean
#                    description: "Признак доступности отправки в 1с"
#                is_pushed_unf_button:
#                    type: boolean
#                    description: "Признак нажали на кнопку в 1с"
#                is_unf_send:
#                    type: boolean
#                    description: "Признак отправили в УНФ"
#                is_paid:
#                    type: boolean
#                    description: "Оплачен"
#                paid_at:
#                    type: string
#                    format: date-time
#                    description: "Дата оплаты"
#                comment:
#                    type: string
#                    description: "Комментарий"
#                cancelled_comment:
#                    type: string
#                    description: "Комментарий к аннулированию аванса"


#        ContractorResponse:
#            type: array
#            items:
#                $ref: "#/components/schemas/ContractorDTO"
#
#        ContractorDTO:
#            type: object
#            properties:
#                id:
#                    type: integer
#                    format: int32
#                    description: "ID контрагента"
#                full_name:
#                    type: string
#                    description: "Наименование"
#                inn:
#                    type: string
#                    description: "ИНН"

        #can use for Messaging?
        CarrierContactDTO:
            type: object
            description: Персональные данные
            required:
                - contractor_id
                - full_name
                - phone
                - email
            properties:
                contractor_id:
                    type: integer
                    format: int64
                    description: "ID"
                full_name:
                    type: string
                    description: "ФИО контактного лица"
                phone_number:
                    type: string
                    pattern: '^9\d{9}$'
                    minLength: 10
                    maxLength: 10
                    example: "9851112233"
                    description: "Номер телефона"
                email:
                    type: string
                    description: "Email-адрес"
                    example: user@example.com
                    format: email
                uuid:
                    type: string
                    description: "Идентификатор контакта"
                    example: 123e4567-e89b-12d3-a456-426655440000
                    format: uuid
                created_at:
                    type: string
                    format: date-time
                    example: 2017-07-21T17:32:28Z
                    description: "Время создания контакта"
...
        isTripAdvanced:
            type: object
            description: "Статус выдачи аванса для Трипа"
            properties:
                isButtonActive:
                    type: boolean
                    description: "Доступность выдачи аванса - когда аванса по трипу нет!"
                contactsNotFound:
                    type: boolean
                    description: "Признак блокировки контрагента - isContractorLock - Используем, если у Контрагента нет контакта авансирования?"
                authorId:
                    type: integer
                    format: int64
                    description: "Id того кто нажал кнопку \"Выдать аванс\""
                firstName:
                    type: string
                    description: "Имя того кто нажал кнопку \"Выдать аванс\""
                lastName:
                    type: string
                    description: "Фамилия того кто нажал кнопку \"Выдать аванс\""
                middleName:
                    type: string
                    description: "Отчество того кто нажал кнопку \"Выдать аванс\""
                createdAt:
                    type: string
                    format: date-time
                    description: "Время и дата запроса аванса"
                isAutoRequested:
                    type: boolean
                    description: "--- isAutoRequested"
...